clean up compiler; we now have room for 32bits indices

+++ r_comp/compiler.h
@@ -89,8 +89,8 @@ private:
     /// In high-level pattern
     bool in_hlp;
     std::vector<std::string> hlp_references;
-    uint8 add_hlp_reference(std::string reference_name);
-    uint8 get_hlp_reference(std::string reference_name);
+    uint32_t add_hlp_reference(std::string reference_name);
+    uint32_t get_hlp_reference(std::string reference_name);
 
 // Utility.
     bool read_nil(RepliStruct *node, uint16 write_index, uint16 &extent_index, bool write);
